+++ crates/biome_analyze/src/syntax.rs
@@ -1,5 +1,4 @@
 use biome_rowan::{AstNode, Language, SyntaxNode, WalkEvent};
-use tracing::debug;
 
 use crate::{
     registry::NodeLanguage, AddVisitor, Phases, QueryKey, QueryMatch, Queryable, ServiceBag,
@@ -11,8 +10,8 @@ use crate::{
 pub struct Ast<N>(pub N);
 
 impl<N> Queryable for Ast<N>
-where
-    N: AstNode + 'static,
+    where
+        N: AstNode + 'static,
 {
     type Input = SyntaxNode<NodeLanguage<N>>;
     type Output = N;
@@ -63,7 +62,6 @@ impl<L: Language + 'static> Visitor for SyntaxVisitor<L> {
     type Language = L;
 
     fn visit(&mut self, event: &WalkEvent<SyntaxNode<Self::Language>>, mut ctx: VisitorContext<L>) {
-        debug!("Node range {:?} and given range {:?}", &event, &ctx.range);
         let node = match event {
             WalkEvent::Enter(node) => node,
             WalkEvent::Leave(node) => {
@@ -94,7 +92,6 @@ impl<L: Language + 'static> Visitor for SyntaxVisitor<L> {
 
 #[cfg(test)]
 mod tests {
-
     use biome_rowan::{
         raw_language::{RawLanguage, RawLanguageKind, RawLanguageRoot, RawSyntaxTreeBuilder},
         AstNode, SyntaxNode,