+++ crates/biome_cli/src/execute/process_file/format.rs
@@ -5,7 +5,7 @@ use crate::execute::process_file::{
 };
 use crate::execute::TraversalMode;
 use biome_diagnostics::{category, DiagnosticExt};
-use biome_service::file_handlers::{AstroFileHandler, VueFileHandler, SVELTE_FENCE};
+use biome_service::file_handlers::{AstroFileHandler, VueFileHandler, SvelteFileHandler};
 use biome_service::workspace::RuleCategories;
 use std::path::Path;
 use std::sync::atomic::Ordering;
@@ -64,40 +64,20 @@ pub(crate) fn format_with_guard<'ctx>(
 
             let mut output = printed.into_code();
 
-            // NOTE: ignoring the
-            if ignore_errors {
+            if ignore_errors || output.is_empty() {
                 return Ok(FileStatus::Ignored);
             }
 
             match workspace_file.as_extension() {
                 Some("astro") => {
-                    if output.is_empty() {
-                        return Ok(FileStatus::Ignored);
-                    }
                     output = AstroFileHandler::output(input.as_str(), output.as_str());
                 }
                 Some("vue") => {
-                    if output.is_empty() {
-                        return Ok(FileStatus::Ignored);
-                    }
                     output = VueFileHandler::output(input.as_str(), output.as_str());
                 }
 
                 Some("svelte") => {
-                    if output.is_empty() {
-                        return Ok(FileStatus::Ignored);
-                    }
-                    if let Some(script) = SVELTE_FENCE
-                        .captures(&input)
-                        .and_then(|captures| captures.name("script"))
-                    {
-                        output = format!(
-                            "{}{}{}",
-                            &input[..script.start()],
-                            output.as_str(),
-                            &input[script.end()..]
-                        );
-                    }
+                    output = SvelteFileHandler::output(input.as_str(), output.as_str());
                 }
                 _ => {}
             }